return formats of apis adjusted, code restructured
